Update zhs.json
t5" : "Intensity Minutes"
does it mean solar intensity minutes? 
No, these are active minutes. So if you do a 30 minute workout with medium intensity, it will show 30.

Alternate Time 1 is defined by Timezone 1, am I right ?
Yes, that is correct. It might make sense to match the words better. _custom_time0 is the setting, f864 is the field that is picked to show it. So instead of alternative time 1, better might be Timezone 1. I leave that up to you.


        "u58": " mb",
My translation is MegaByte (MB) about storage. Is it right? 
My translation is MegaByte (MB) about storage. Is it right?
No, this is MegaBar, used for air pressure.

"g7"  to "g19", g23, g24, g25, g28
It looks like the AI added these in to fill the gap. I removed them, since they are not necessary
--.> you did not remove yet. so i removed.

   "labels" : {

Correct. These are the labels that are used. I did not translate them, AI will definitely make a mess of it. It would be great to have these included in other languages as well, but this will be a manual process. 
--> i copy the line 722 to line 755 from ENG to ZHS. Please check if it is correct (i'm not very clear about {{...}} matches.  I know little about programming languages.