Time Format (Time) vs Time Format (Data)
 "t2": "Time Format (Time)",  i translated to Time Format (the main time)
"t3": "Time Format (Data)", i translated to Time Format (Other times) (which means sunset time, dusk time, alternative time, etc.)

The translations may not be guaranteed to match the original text 100%, but they should be easier for the user to understand. 100% direct translations of the original text may not be understandable.
Please take a look at this translation to see if it's basically the same as what you're saying.