Refine the translation (line 1 to line 58)
I further refined the translation based on Garmin Owner's Manual in English  and in Simplified Chinese and the context from Garmin Website. 
The first step to translate is to understand the English 100% clearly. 
To understand each term, I have to try these data fields on watch face.  It really takes time as I want to make it perfectly.