Update eng.json - PLEASE REVIEW CAREFULLY IF YOU ACCEPT OR NOT
Usually I should not make any changes on the original eng.json as a translator. 
While as a user, i do not think the terms used are very clear.  
for some settings, it has confusions and conflicts. 
for example, "_dc": "Date Color (row 4)", I prefer to set row 4 to steps, then the setting of date color is not correct. 

Because you have nominated the data fields to row 1 - 4 at the color settings, i prefer to use row 1-4 for the data filed settings. 

To help understanding, i added (top row) behind row 1. 
If possible, please add an illustration photo in next update, showing what are the positions of Row 1-4 and bottom text. 

    "_ic": "Top Icons time Color (row 1)", I think "time" is not required.