Update eng.json of enduro2 - PLEASE REVIEW CAREFULLY
  "_amc": "Center Right Bottom Color / Center Right label Color (bottom)",
Is it necessary to use two names?
I think we could use one of the two names. 
beside, there is 
       "_l9": "9. Center Right Label (bottom)",
So i rename to Center Right label Color (bottom)


"_lc": "Label Color",
As per discsssion, it applies to icons only, or if the text label is above or below the data.
For Enduro 2, There are no labels above or below the data. So we rename it to Icon color  or Icon label color. 


        "_tc": "Data Color",
        "_tf": "Text Font",
        "_tfs": "Text Font size",
text font applies to both data and text label left/right of the data. so i do the rename. 

        "_tc": "Data and Text Label Color",
        "_tf": "Data and Text Label Font",
        "_tfs": "Data and Text Label Font size",


"_scc": "Seconds Color",
i rename it to:
Center Right Data Color (top, default - seconds)


 "_ttfs": "Top/Bottom Text Font"
It is Font Size, not font types, such as bold, irregular, italic...
So i rename it to Top/Bottom Text Font Size