Update eng.json - PLEASE CHECK CAREFULLY IF THE CHANGES ARE CORRECT
"_cpbs" : "Same as Left Top Progress Bar",
Is it a type mistake? I think it should be same as top progress bar


        "_o2": "Field X Offsets (Top, Left, Right, Bottom, Progress 1, Progress 2) Default: 20",
        "_o3": "Field Y Offsets (Top, Left Text, Right Text, Bottom, Progress 1, Progress 2, Left Label, Right Label) Default: 20",
I guess Progress 1/2 actually means Top/bottom Progress.
The progress bar names at the start of the settings and in the illustration photo are: top/bottom progress.
So i think it brings a little confusion to use progress 1/2 at the end of the settings.